+++ app/vehicles/[vehicle]/page.tsx
@@ -85,7 +85,7 @@ export default async function VehiclePage({
       </ul>
 
       <p className={styles.more}>
-        <Link href="/for-parents">家長指南</Link>
+        <Link href="/for-parents">親子指南</Link>
       </p>
 
       <SiteFooter />